django_dcmn/requirements.txt
gunicorn
django-cors-headers
django-redis
django-anymail

celery[redis]

asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.2
djangorestframework==3.16.0
idna==3.10
psycopg[binary]>=3.1.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
stripe==12.0.0
typing_extensions==4.13.1
urllib3==2.3.0
whitenoise==6.9.0
